feat: 유저 검색 api 결과 업데이트사항 반영
