+++ bookduck/src/main/java/com/mmc/bookduck/domain/user/dto/common/UserUnitDto.java
@@ -1,16 +1,22 @@
 package com.mmc.bookduck.domain.user.dto.common;
 
+import com.mmc.bookduck.domain.item.dto.common.ItemEquippedUnitDto;
 import com.mmc.bookduck.domain.user.entity.User;
+
+import java.util.List;
 
 public record UserUnitDto(
         Long userId,
-        String nickname
-        // TODO: 캐릭터 이미지 추가해야
+        String nickname,
+        boolean isFriend,
+        List<ItemEquippedUnitDto> itemEquipped
 ) {
-    public static UserUnitDto from (User user) {
+    public static UserUnitDto from (User user, List<ItemEquippedUnitDto> itemEquipped, boolean isFriend) {
         return new UserUnitDto(
                 user.getUserId(),
-                user.getNickname()
+                user.getNickname(),
+                isFriend,
+                itemEquipped
         );
     }
 }
