+++ bookduck/src/main/java/com/mmc/bookduck/domain/user/service/UserService.java
@@ -1,23 +1,15 @@
 package com.mmc.bookduck.domain.user.service;
 
-import com.mmc.bookduck.domain.user.dto.common.UserUnitDto;
-import com.mmc.bookduck.domain.user.dto.response.UserInfoResponseDto;
-import com.mmc.bookduck.domain.user.dto.response.UserSearchResponseDto;
 import com.mmc.bookduck.domain.user.entity.User;
-import com.mmc.bookduck.domain.user.entity.UserGrowth;
 import com.mmc.bookduck.domain.user.entity.UserStatus;
 import com.mmc.bookduck.domain.user.repository.UserRepository;
 import com.mmc.bookduck.global.exception.CustomException;
 import com.mmc.bookduck.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import static com.mmc.bookduck.global.common.EscapeSpecialCharactersService.escapeSpecialCharacters;
 
 @Service
 @RequiredArgsConstructor
@@ -59,21 +51,6 @@ public class UserService {
     @Transactional
     public User saveUser(User user){
         return userRepository.save(user);
-    }
-
-    // 유저 검색
-    public UserSearchResponseDto searchUsers(String keyword, Pageable pageable) {
-        // 키워드의 이스케이프 처리
-        String escapedWord = escapeSpecialCharacters(keyword);
-        Page<User> userPage = getSearchedUserPage(escapedWord, pageable);
-
-        Page<UserUnitDto> userUnitDtoPage = userPage.map(UserUnitDto::from);
-        return UserSearchResponseDto.from(userUnitDtoPage);
-    }
-
-    @Transactional(readOnly = true)
-    public Page<User> getSearchedUserPage(String keyword, Pageable pageable) {
-        return userRepository.searchAllByNicknameContaining(keyword, pageable);
     }
 
     @Transactional(readOnly = true)
